# openapi.yaml
openapi: 3.0.0
info:
  title: NethCTI Middleware API
  description: |
    NethCTI Middleware provides authentication, 2FA, phonebook management, 
    and integration with NethVoice/NethCTI systems.
  version: 1.0.0
  contact:
    name: Nethesis
    url: https://www.nethesis.it
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://middleware.example.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /login endpoint
    superAdminAuth:
      type: http
      scheme: bearer
      description: Super admin bearer token for administrative endpoints
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
...
    PhonebookImportResponse:
      type: object
      properties:
        message:
          type: string
        total_rows:
          type: integer
          description: Total rows in the CSV
        imported_rows:
          type: integer
          description: Successfully imported rows
        failed_rows:
          type: integer
          description: Failed rows
        skipped_rows:
          type: integer
          description: Skipped rows (validation errors)
        error_messages:
          type: array
          items:
            type: string
          description: List of error messages encountered during import
...
paths:
...
  /phonebook/import:
    post:
      summary: Import phonebook entries from CSV file
      operationId: importPhonebookCSV
      tags:
        - Phonebook
      security:
        - bearerAuth: []
      description: |
        Import phonebook entries from a CSV file. The CSV must have a 'name' column (required).
        
        Supported optional columns (case-insensitive):
        - name (required): Contact name
        - type (optional): Entry type - must be 'private' or 'public' (defaults to 'private')
        - workphone (optional): Work phone number
        - homephone (optional): Home phone number
        - cellphone (optional): Cell phone number
        - workemail (optional): Work email address
        - homeemail (optional): Home email address
        - company (optional): Company name
        - title (optional): Job title
        - fax (optional): Fax number
        - notes (optional): Notes/description
        - homestreet (optional): Home street address
        - homepob (optional): Home P.O. box
        - homecity (optional): Home city
        - homeprovince (optional): Home province/state
        - homepostalcode (optional): Home postal code
        - homecountry (optional): Home country
        - workstreet (optional): Work street address
        - workpob (optional): Work P.O. box
        - workcity (optional): Work city
        - workprovince (optional): Work province/state
        - workpostalcode (optional): Work postal code
        - workcountry (optional): Work country
        - url (optional): Website URL
        - extension (optional): Phone extension
        - speeddial_num (optional): Speed dial number

        Validation rules:
        - Rows with empty 'name' field are skipped
        - Rows with invalid 'type' (not 'private' or 'public') are skipped
        - All other fields are optional and can be empty
        - Type values are case-insensitive and normalized to lowercase

        Example CSV:
        ```
        name,type,workphone,cellphone,company
        John Doe,private,+1234567890,,Example Corp
        Jane Smith,public,,+0987654321,
        Bob Johnson,,,+1111111111,Tech Services
        ```
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file (text/csv or text/plain)
      responses:
        '200':
          description: Import completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhonebookImportResponse'
        '400':
          description: Invalid file or CSV format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '500':
          description: Server error during import
...